move feature to EE
